Get provider schema properly (#34)
* Mention Terraform provider version

This addition in the provider info tells the Pulumi Bridge that the Go module of the provider is now at v2. This ensures that we fetch the provider properly without any silent errors.

* Re-generate data based on resource config

This was regenerated using `make build_sdks`

* Fix NPM package name in documentation

This is a known issue when generating the Node SDK. `@pulumi/onepassword` should be `@1Password/pulumi-onepassword`.

* Remove ignored files

The files are now properly generated, so they can be removed from ignored files to check to ensure they're up to date.